Remove shared code persistence module
